requirements.txt
streamlit
pillow
cv2
tensorflow
numpy
pandas
matplotlib
